Fix invalid :pre syntax
The code being fixed in this commit does not use proper syntax
to specify :pre conditions. As a result, when any of the predicate
functions are provided with the `response` parameter that is not a map,
instead of throwing an assertion error, execution reaches
`(:status response)` whereupon it produces `NullPointerException`.